Clean up values and configs:
Rename the BRANCHOUT_DIRECTORY to PROJECTION directory this make its much clearer what it means. The place we checkout a projection is different than the state of the usage of the projection.

Make sure that all values in the Branchoutfile are prefixed with BRANCHOUT_ just like the configs.

Only source the branchout-configuration when its used, make debugging a lot easier